verifyEmail changes
