+++ app/package.json
@@ -28,5 +28,10 @@
     "eslint-config-next": "14.1.3",
     "postcss": "^8",
     "tailwindcss": "^3.3.0"
+  },
+  "browser": {
+    "fs": false,
+    "path": false,
+    "os": false
   }
 }
