+++ app/pages/api/auth/signup.js
@@ -2,7 +2,7 @@ import dbConnect from "@/lib/dbConnect";
 import bcrypt from "bcryptjs";
 const User = require("@/models/user.js");
 import { z } from "zod";
-import nodemailer from "nodemailer";
+import sendVerificationEmail from "@/Helper/sendVerificationEmail";
 
 const passwordSchema = z
   .string()
@@ -83,80 +83,9 @@ const handler = async (req, res) => {
     const result = await user.save();
     res.json({ success: true });
     console.log(user);
-    sendVerificationEmail(user.email, user.firstName,activation_token);
+    sendVerificationEmail(user.email, user.firstName, activation_token);
   } catch (err) {
     res.status(500).json({ success: false, message: "Server Error" });
-  }
-
-};
-
-const sendVerificationEmail = async (userEmail, firstName,activation_token) => {
-  try {
-    let transporter = nodemailer.createTransport({
-      service: "gmail",
-      auth: {
-        user: "rajancoding597@gmail.com",
-        pass: "funzmtuwnobqjvya", // Use a secure method to store and access passwords
-      },
-    });
-
-    let info = await transporter.sendMail({
-      from: '"Connect" <rajancoding597@gmail.com>',
-      to: userEmail,
-      subject: "Verify Your Email Address",
-      text: `Hello ${firstName},\nPlease verify your email address by clicking on the following link: [Verification Link Here]`,
-      html: `
-      <html>
-      <head>
-        <style>
-          body {
-            font-family: Arial, sans-serif;
-            background-color: #f7f7f7;
-            margin: 0;
-            padding: 0;
-          }
-          .container {
-            max-width: 600px;
-            margin: 0 auto;
-            padding: 20px;
-            background-color: #fff;
-            border-radius: 5px;
-            box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-          }
-          h1 {
-            color: #333;
-          }
-          p {
-            color: #555;
-          }
-          .button {
-            display: inline-block;
-            background-color: #007bff;
-            color: #fff;
-            text-decoration: none;
-            padding: 10px 20px;
-            border-radius: 5px;
-            transition: background-color 0.3s;
-          }
-          .button:hover {
-            background-color: #0056b3;
-          }
-        </style>
-      </head>
-      <body>
-        <div class="container">
-          <h1>Hello ${firstName},</h1>
-          <p>Please verify your email address by clicking on the following button:</p>
-          <a class="button" href="https://localhost:3000/api/auth/verifyEmail?ActivationToken='${activation_token}'">Verify Email</a>
-        </div>
-      </body>
-      </html>
-      `,
-    });
-
-    console.log("Message sent: %s", info.messageId);
-  } catch (err) {
-    console.error(err); // Log error for debugging
   }
 };
